+++ frontend/types/messages.d.ts
@@ -18,6 +18,8 @@ type AppMessages = typeof import("../messages/en.json") & {
   Plugin_pushover: typeof import("../plugins/pushover/locales/en.json");
   Plugin_discord: typeof import("../plugins/discord/locales/en.json");
   Plugin_telegram: typeof import("../plugins/telegram/locales/en.json");
+  Plugin_gotify: typeof import("../plugins/gotify/locales/en.json");
+  Plugin_pagerduty: typeof import("../plugins/pagerduty/locales/en.json");
 };
 
 declare module "next-intl" {